git-svn-id: svn://192.168.1.10/public/CoffeeMud@9663 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/WebMacros/AbilityData.java
@@ -371,7 +371,7 @@ public class AbilityData extends StdWebMacro
 					int o=CMath.s_int(old);
 					str.append("<OPTION VALUE=\"-1\""+((o==-1)?" SELECTED":"")+">Use Default");
 					str.append("<OPTION VALUE=\"0\""+((o==0)?" SELECTED":"")+">None (free skill)");
-					str.append("<OPTION VALUE=\"\""+(((o>0)&&(o<Integer.MAX_VALUE-101))?" SELECTED":"")+"\">Custom Value");
+					str.append("<OPTION VALUE=\"\""+(((o>0)&&(o<Short.MAX_VALUE-101))?" SELECTED":"")+"\">Custom Value");
 					str.append("<OPTION VALUE=\""+Short.MAX_VALUE+"\""+((o==Short.MAX_VALUE)?" SELECTED":"")+">All Mana");
 					for(int v=Short.MAX_VALUE-5;v>=Short.MAX_VALUE-95;v-=5) {
 						str.append("<OPTION VALUE=\""+v+"\""+(((o>(v-5))&&(o<=v))?" SELECTED":"")+">"+(Short.MAX_VALUE-v)+"%");
